(BSR) feat(homepageListFixture): rename fixture
