+++ src/features/home/api/useHomepageData.native.test.ts
@@ -1,7 +1,7 @@
 /* eslint-disable local-rules/no-react-query-provider-hoc */
 import { rest } from 'msw'
 
-import { homepageList } from 'features/home/fixtures/homepageList.fixtures'
+import { homepageList } from 'features/home/fixtures/homepageList.fixture'
 import { Homepage } from 'features/home/types'
 import { BASE_URL, PARAMS } from 'libs/contentful/fetchHomepageNatifContent'
 import { homepageEntriesAPIResponse } from 'libs/contentful/fixtures/homepageEntriesAPIResponse'
